rename: component stepper, cart button 이름 변경
