+++ src/views/Cart/components/CartButtonWithIcon/index.tsx
@@ -1,0 +1,1 @@
+export { default as CartButtonWithIcon } from './CartButtonWithIcon';
