--- src/views/Cart/components/CartQuantityField/index.tsx
@@ -1,1 +1,0 @@
-export { default as CartQuantityField } from './CartQuantityField';
