+++ src/views/Cart/components/CartStepper/index.tsx
@@ -1,0 +1,1 @@
+export { default as CartStepper } from './CartStepper';
