--- src/views/Cart/components/CartStepperWithIcon/index.tsx
@@ -1,1 +1,0 @@
-export { default as CartStepperWithIcon } from './CartStepperWithIcon';
